Improvements to provider type selection
Handles selection of T Levels and how  that is commited following the choice of available T Levels for the provider. Works in a way that replicates what real world interaction should be. This incorporates the selection of T Levels for provider type, and then no selection of associated T Levels and returning to the provider details page.

Mini provider navigation updates according to selected provider types.

Main nav 'Home' link updated to 'Halpdesk dashboard'